# src/main/resources/swagger/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:10080/backend
    description: Backend
...
paths:
...
  /api/v1/fire-weather-indexes/{id}:
    parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
          format: uuid
          description: ID КПО
    get:
      tags:
        - FWIController
      summary: Получение конкретной записи
      description: Получение конкретной записи об КПО
      operationId: getById
      responses:
        '200':
          description: Успешный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FWIResponseDto'
        '400':
          description: Отсутствие id в запросе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Отсутствует параметр id в запросе
                    code: 400
        '401':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Пользователь не авторизован
                    code: 401
        '404':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: КПО не был найден в БД
                  value:
                    message: Запись о КПО 1993dbee-eb94-46cb-aa9d-5b166b1c6659 нет в системе
                    code: 404
        '500':
          description: Ошибка на сервере
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                example1:
                  description: Пример ответа на 500
                  value:
                    message: Произошла ошибка на стороне сервера
                    code: 500
    delete:
      tags:
        - FWIController
      summary: Удалить запись
      description: Удаление записи по КПО
      operationId: deleteById
      responses:
        '200':
          description: Успешное выполнение запроса
        '400':
          description: Отсутствие id в запросе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Отсутствует параметр id в запросе
                    code: 400
                example2:
                  description: Пример ответа 2
                  value:
                    message: КППО связан с другими записями
                    code: 400
        '401':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Пользователь не авторизован
                    code: 401
        '404':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: КПО не был найден в БД
                  value:
                    message: Запись о КПО 1993dbee-eb94-46cb-aa9d-5b166b1c6659 нет в системе
                    code: 404
        '500':
          description: Ошибка на сервере
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                example1:
                  description: Пример ответа на 500
                  value:
                    message: Произошла ошибка на стороне сервера
                    code: 500
  /api/v1/precipitation-records:
    get:
      tags:
        - PrecipitationRecordController
      summary: Получить записи по таблице осадков
      description: Получение всех записей таблицы осадков
      operationId: getPrecipitationRecords
      responses:
        '200':
          description: Успешное получение всех записей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PrecipitationRecordResponseDto'
        '401':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Пользователь не авторизован
                    code: 401
        '500':
          description: Ошибка на сервере
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                example1:
                  description: Пример ответа на 500
                  value:
                    message: Произошла ошибка на стороне сервера
                    code: 500
    post:
      tags:
        - PrecipitationRecordController
      summary: Добавить запись по таблице осадков
      description: Добавление новой записи по природному осадку в БД
      operationId: addPrecipitationRecord
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrecipitationRecordCreateDto'
      responses:
        '201':
          description: Успешное добавление записи
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrecipitationRecordResponseDto'
        '400':
          description: Ошибка в составлении запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                example1:
                  description: Пример неправильного запроса
                  value:
                    message: Минимальная граница больше максимальной
                    code: 400
                example2:
                  description: Пример неправильного запроса
                  value:
                    message: Отрицательное значение в кол-ве осадков
                    code: 400
        '401':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Пользователь не авторизован
                    code: 401
        '403':
          description: Пользователю не хватает прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Недостаточно прав для выполнения операции
                    code: 403
        '500':
          description: Ошибка на сервере
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                example1:
                  description: Пример ответа на 500
                  value:
                    message: Произошла ошибка на стороне сервера
                    code: 500
    patch:
      tags:
        - PrecipitationRecordController
      summary: Обновить запись по осадкам
      description: Измененение минимальной, максимальной границы и кол-во осадков
      operationId: patchRecord
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrecipitationRecordUpdateDto'
      responses:
        '200':
          description: Успешное обновление записи
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrecipitationRecordResponseDto'
        '400':
          description: Ошибка в составлении запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                example1:
                  description: Пример неправильного запроса
                  value:
                    message: Минимальная граница больше максимальной
                    code: 400
                example2:
                  description: Пример неправильного запроса
                  value:
                    message: Отрицательное значение в кол-ве осадков
                    code: 400
        '401':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Пользователь не авторизован
                    code: 401
        '403':
          description: Пользователю не хватает прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Недостаточно прав для выполнения операции
                    code: 403
        '500':
          description: Ошибка на сервере
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                example1:
                  description: Пример ответа на 500
                  value:
                    message: Произошла ошибка на стороне сервера
                    code: 500
  /api/v1/precipitation-records/{id}:
    parameters:
      - name: id
        required: true
        in: path
        schema:
          type: integer
          format: int64
          description: ID записи таблицы осадков
    get:
      tags:
        - PrecipitationRecordController
      summary: Получение конкретной записи
      description: Всё равно меня никто не прочитает
      operationId: getRecordById
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrecipitationRecordResponseDto'
        '401':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Пользователь не авторизован
                    code: 401
        '500':
          description: Ошибка на сервере
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                example1:
                  description: Пример ответа на 500
                  value:
                    message: Произошла ошибка на стороне сервера
                    code: 500
    delete:
      tags:
        - PrecipitationRecordController
      summary: Удалить запись
      description: Удаление конкретной записи по id
      operationId: deleteRecord
      responses:
        '200':
          description: Успешное выполнение запроса
        '400':
          description: Отсутствие id в запросе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Отсутствует параметр id в запросе
                    code: 400
        '401':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Пользователь не авторизован
                    code: 401
        '404':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: КПО не был найден в БД
                  value:
                    message: Запись об осадках с id 1 нет в системе
                    code: 404
        '500':
          description: Ошибка на сервере
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                example1:
                  description: Пример ответа на 500
                  value:
                    message: Произошла ошибка на стороне сервера
                    code: 500
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiExceptionDto:
      type: object
      description: ДТО для ответа с ошибкой
      required:
        - code
        - message
      properties:
        message:
          type: string
          description: Сообщение с ошибкой
        code:
          type: integer
          format: int32
    # DTO для погодных параметров
...
    FWIResponseDto:
      description: DTO представление КПО в приложении
      properties:
        id:
          type: string
          format: uuid
          example: 12efb6fa-72b7-455e-87c2-872d24d1f22d
        name:
          type: string
          example: КПО_Нестерова_1
        minValue:
          type: number
          description: нижняя граница горммости
          minimum: 0
          example: 0
        maxValue:
          type: number
          description: верхняя граница горимости
          minimum: 0
          example: 1000
        fireDangerName:
          type: string
          description: Имя уровня пожарной угрозы
          example: Незначительный
        ffwiName:
          type: string
          description: Название КППО
          example: КППО_Нестерова
...
    PrecipitationRecordResponseDto:
      properties:
        id:
          type: integer
          format: int64
          example: 1
        minValue:
          type: number
          description: Минимальная граница горимости
          example: 0
        maxValue:
          type: number
          description: Максимальная граница горимости
          example: 100000.0
        value:
          type: integer
          description: Значение осадков в мм
          example: 3
        ffwiName:
          type: string
          description: Название КППО
          example: КППО_Нестерова
        ffwId:
          type: string
          format: uuid
          description: ID КППО
          example: 97b4e98f-6829-446a-aea7-8e7be17d56c9
    PrecipitationRecordCreateDto:
      required:
        - minValue
        - maxValue
        - value
        - ffwi_id
      properties:
        minValue:
          type: number
          description: Минимальная граница горимости
          example: 0
        maxValue:
          type: number
          description: Максимальная граница горимости
          example: 100000.0
        value:
          type: integer
          description: Значение осадков в мм
          example: 3
        ffwi_id:
          type: string
          format: uuid
          description: ID КППО
          example: 97b4e98f-6829-446a-aea7-8e7be17d56c9
    PrecipitationRecordUpdateDto:
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
          example: 1
        minValue:
          type: number
          description: Минимальная граница горимости
          example: 0
        maxValue:
          type: number
          description: Максимальная граница горимости
          example: 100000.0
        value:
          type: integer
          description: Значение осадков в мм
          example: 3

    # DTO для пользователя
...